+++ frontend/src/component/Tables/PMC/PMCTable.js
@@ -109,8 +109,11 @@ const PMCTable = ({ modalHandler, refresh,isOpen }) => {
     csvExporter.generateCsv(rows.map((row) => row.original));
   };
   const delteHandler = async (id) => {
-    const data = await axios.delete(`/api/v1/pmc/delete/${id}`);
-    fetchPMC();
+    // eslint-disable-next-line no-restricted-globals
+    if(window.confirm("Are you sure ?")){
+      const data = await axios.delete(`/api/v1/pmc/delete/${id}`);
+      fetchPMC();
+    }
   }
 
   const fetchPMC = async () => {
